update AWS SDK to v2
AWS SDK v2 has more advanced detection of the running environment (ECS,
EC2, etc) and it is recommended to pull credentials from the default
provider chain.

https://docs.aws.amazon.com/sdk-for-java/v1/developer-guide/credentials.html

All functionality should be maintained.

Note the explicit expiration of presigned S3 object requsts. In the V1
SDK, not setting an expiration resulted in a default 15 minute
expiration time. It is not possible to set no expiration time for
presigned requests.

https://github.com/aws/aws-sdk-java/blob/master/aws-java-sdk-s3/src/main/java/com/amazonaws/services/s3/AmazonS3Client.java#L3469